Break classes into separate files
