+++ lib/game.rb
@@ -1,0 +1,12 @@
+require 'game_state'
+
+class Game
+  def initialize
+    @score = 0
+    @player = Player.new
+  end
+
+  def generate_map(size)
+    map_array = []
+  end
+end
