+++ lib/game_state.rb
@@ -1,5 +1,3 @@
-# GameState
-
 class GameState
 
 end
